refactor: move ConsumerPurposeTemplateLinkedResourceTab to components/shared (PIN-10158)
The tab is consumed by both ConsumerPurposeTemplateDetailsPage and
ConsumerPurposeTemplateCatalogDetailsPage. Living under one page's components
directory required the sibling page to cross-import. Move it next to
ConsumerPurposeTemplateLinkedResourceTable.
