+++ src/pages/ConsumerPurposeTemplateDetailsPage/ConsumerPurposeTemplateDetails.page.tsx
@@ -8,7 +8,7 @@ import { ConsumerPurposeTemplateDetailsTab } from './components/ConsumerPurposeT
 import { TabContext, TabList, TabPanel } from '@mui/lab'
 import { Tab } from '@mui/material'
 import { useActiveTab } from '@/hooks/useActiveTab'
-import { ConsumerPurposeTemplateLinkedResourceTab } from './components/ConsumerPurposeTemplateLinkedResourceTab'
+import { ConsumerPurposeTemplateLinkedResourceTab } from '@/components/shared/PurposeTemplate/PurposeTemplateLinkedResourceTab/ConsumerPurposeTemplateLinkedResourceTab'
 import { PurposeTemplateRiskAnalysisTab } from '@/components/shared/PurposeTemplate/PurposeTemplateRiskAnalysisTab'
 import { TenantHooks } from '@/api/tenant/tenant.hooks'
